fix 1 parameter

+++ libft/ft_atoi.c
@@ -21,7 +21,7 @@ static int	ft_isspace(char c)
 		return (0);
 }
 
-int	ft_atoi(const char *str)
+long	ft_atoi(const char *str)
 {
 	long	nbr;
 	long	sign;
